Fix post member_id

+++ src/main/java/com/realtimechat/client/service/PostService.java
@@ -1,5 +1,6 @@
 package com.realtimechat.client.service;
 
+
 import com.realtimechat.client.domain.Member;
 import com.realtimechat.client.domain.Post;
 import com.realtimechat.client.dto.request.PostRequestDto;
@@ -36,20 +37,33 @@ public class PostService {
     
     // update
     @Transactional
-    public Integer update(Integer id, PostRequestDto requestDto) {
+    public String update(Integer id, PostRequestDto requestDto) {
+        String message = "fail";
         Post post = postRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("해당 게시글이 없습니다."));
 
-        post.update(requestDto.getContent());
-        
-        return id;
+        // 작성자만 수정 가능 
+        if (postRepository.findById(id).get().getMember().getId().equals(requestDto.getMember().getId())) {
+            post.update(requestDto.getContent());
+            message = "success";
+        } 
+
+        return message;
 
     }
 
     // delete
     @Transactional
-    public void delete(Integer id) {
-        Post post = postRepository.findById(id).get();
-        postRepository.delete(post);
+    public String delete(Integer id, Member member) {
+        String message = "fail";
+        Post post = postRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("해당 게시글이 없습니다."));
+        
+        if (post.getMember().getId().equals(member.getId())) { // 작성자만 삭제 가능
+            postRepository.delete(post);
+            message = "success";
+        }
+
+        return message;
+        
     }
 
 }
